React 16
This commit introduces React v16. Most of the changes involve using the
new prop-types module that was pulled out of React. Also there was some
changes to the use of refs, since React no longer lets you definie
refs as a string and using them as this.refs.  I also added webpack for
serving up the baxley example.

More work needs to be done to serve up the other examples, and also to
bundle things for npm.